+++ src/engine/wasmEngine/wasm/src/asc/initImages.ts
@@ -6,7 +6,7 @@ import { PTR_T, SIZE_T } from './memUtils';
 
 function initImages(): SArray<BitImage> {
   const bitImages = newSArray<BitImage>(numImages);
-  for (let i: u32 = 0; i < numImages; i++) {
+  for (let i: usize = 0; i < numImages; i++) {
     const image = bitImages.at(i);
     image.init(i);
   }
